Updated to latest docusaurus 2.2.0
