--- src/theme/Layout/styles.css
@@ -1,27 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-html,
-body {
-  height: 100%;
-}
-
-#__docusaurus {
-  min-height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.main-wrapper {
-  flex: 1 0 auto;
-}
-
-/* Docusaurus-specific utility classes */
-
-.docusaurus-mt-lg {
-  margin-top: 3rem;
-}
